chore: major codebase update
- Add unified container architecture and lazy loading system
- Add transfer/export functionality with progress tracking
- Add document viewers (PDF, spreadsheet, text, hex)
- Add case documents panel with COC form detection
- Add activity panel for session logging
- Add performance panel for system stats
- Add settings panel with preferences
- Add theme system with CSS variables
- Add keyboard shortcuts and command palette
- Refactor EvidenceTree with adapters for different container types
- Add nested container support for archives within containers
- Update report generation with templates
- Add search functionality across containers
- Update Rust backend with new commands and viewers
- Add project file v2 format with caching support
- Update build artifacts and coverage reports
- Clean up deprecated code and documentation
- Add GitHub workflows and VS Code settings

+++ src-tauri/src/archive/tar.rs
@@ -190,7 +190,7 @@ fn read_tar_archive<R: Read>(reader: R, archive_path: &str) -> Result<Vec<Archiv
             .unwrap_or_default();
         
         // Get checksum
-        let cksum = header.cksum().unwrap_or(0) as u32;
+        let cksum = header.cksum().unwrap_or(0);
         
         // Determine compression method description
         let compression = TarCompression::from_extension(archive_path);
